requirements.txt
requests>=2.28.0
pandas>=1.5.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0